Fixing scripts

+++ 2-do_deploy_web_static.py
@@ -4,6 +4,7 @@
 """
 from fabric.api import *
 from fabric.operations import run, put, sudo
+import os
 env.hosts = ['66.70.184.249', '54.210.138.75']
 
 
@@ -11,23 +12,24 @@ def do_deploy(archive_path):
     """
         using fabric to distribute archive
     """
+    if os.path.isfile(archive_path) is False:
+        return False
     try:
-        archive = archive_path.split("/")
-        archive = archive[1]
+        archive = archive_path.split("/")[-1]
+        path = "/data/web_static/releases"
         put("{}".format(archive_path), "/tmp/{}".format(archive))
         folder = archive.split(".")
-        run("mkdir -p /data/web_static/releases/{}/".format(folder[0]))
+        run("mkdir -p {}/{}/".format(path, folder[0]))
         new_archive = '.'.join(folder)
-        run("tar -xzf /tmp/{} -C /data/web_static/releases/{}/"
-            .format(new_archive, folder[0]))
-        run("mv /data/web_static/releases/{}/web_static/*"
-            " /data/web_static/releases/{}"
-            .format(folder[0], folder[0]))
-        run("rm -rf /data/web_static/releases/{}/web_static".format(folder[0]))
+        run("tar -xzf /tmp/{} -C {}/{}/"
+            .format(new_archive, path, folder[0]))
         run("rm /tmp/{}".format(archive))
+        run("mv {}/{}/web_static/* {}/{}/"
+            .format(path, folder[0], path, folder[0]))
+        run("rm -rf {}/{}/web_static".format(path, folder[0]))
         run("rm -rf /data/web_static/current")
-        run("ln -sf /data/web_static/releases/{} /data/web_static/current"
-            .format(folder[0]))
+        run("ln -sf {}/{} /data/web_static/current"
+            .format(path, folder[0]))
         return True
     except:
         return False
